+++ flsim/__init__.py
@@ -5,8 +5,10 @@ __all__ = [
     "partition",
     "node",
     "aggregator",
-    "aggregation",
-    "eval",
+    "federated_aggregation",
+    "evaluation",
+    "incentive_strategy",
+    "contracts",
     "attacks",
     "runner",
     "malicious_detection",
