+++ flsim/aggregator.py
@@ -1,3 +1,3 @@
-from .aggregation.aggregator import Aggregator
+from .federated_aggregation.aggregator import Aggregator
 
 __all__ = ["Aggregator"]
