+++ flsim/contracts/__init__.py
@@ -1,0 +1,5 @@
+from .base import BaseContract
+from .ours import OurContract
+from .baseline1 import Baseline1Contract
+
+__all__ = ["BaseContract", "OurContract", "Baseline1Contract"]
