+++ flsim/contracts/baseline1.py
@@ -1,0 +1,9 @@
+"""Baseline contract for comparisons."""
+from .base import BaseContract
+
+
+class Baseline1Contract(BaseContract):
+    """Simple baseline contract variant."""
+    pass
+
+__all__ = ["Baseline1Contract"]
