+++ flsim/evaluation/__init__.py
@@ -1,0 +1,17 @@
+from .modeleval import (
+    evaluate_global,
+    evaluate_state_dict,
+    evaluate_many_state_dicts,
+    evaluate_reconstructed_single,
+    evaluate_reconstructed_batch,
+    reconstruct_state,
+)
+
+__all__ = [
+    "evaluate_global",
+    "evaluate_state_dict",
+    "evaluate_many_state_dicts",
+    "evaluate_reconstructed_single",
+    "evaluate_reconstructed_batch",
+    "reconstruct_state",
+]
