+++ flsim/federated_aggregation/__init__.py
@@ -1,8 +1,8 @@
 """Aggregation utilities and strategies."""
 
 from .aggregator import Aggregator
-from .fedavg import Strategy as FedAvg
-from .committee_fedavg import Strategy as CommitteeFedAvg
+from .FedAvg import Strategy as FedAvg
+from .CommitteeFedAvg import Strategy as CommitteeFedAvg
 from .trimmed_mean import Strategy as TrimmedMean
 
 __all__ = [
