+++ flsim/models/__init__.py
@@ -1,0 +1,3 @@
+from .model import LinearMNIST
+
+__all__ = ["LinearMNIST"]
